% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JMstateModel.R
\name{JMstateModel}
\alias{JMstateModel}
\title{Title}
\usage{
JMstateModel(
  lmeObject,
  survObject,
  timeVar,
  parameterization = c("value", "slope", "both"),
  method = c("weibull-PH-aGH", "weibull-PH-GH", "weibull-AFT-aGH", "weibull-AFT-GH",
    "piecewise-PH-aGH", "piecewise-PH-GH", "Cox-PH-aGH", "Cox-PH-GH", "spline-PH-aGH",
    "spline-PH-GH", "ch-Laplace"),
  interFact = NULL,
  derivForm = NULL,
  lag = 0,
  scaleWB = NULL,
  CompRisk = FALSE,
  init = NULL,
  control = list(),
  Mstate = FALSE,
  data.Mstate = NULL,
  ID.Mstate = NULL,
  init.type.ranef = "mean",
  b.postVar.opt = FALSE,
  transform.value = NULL,
  gr.transform.value = NULL,
  H.transform.value = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
